+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/JwsAlgorithm.kt
@@ -219,12 +219,12 @@ fun MessageAuthenticationCode.toJwsAlgorithm(): KmmResult<JwsAlgorithm> = catchi
 
 /** Tries to find a matching JWS algorithm*/
 fun SpecializedDataIntegrityAlgorithm.toJwsAlgorithm() =
-    this.algorithm.toJwsAlgorithm()
+    this.algorithm?.toJwsAlgorithm()?: KmmResult.failure(UnsupportedCryptoException("Unsupported algorithm: ${this::class.simpleName}"))
 
 /** Tries to find a matching JWS algorithm.*/
 fun SpecializedMessageAuthenticationCode.toJwsAlgorithm() =
-    this.algorithm.toJwsAlgorithm()
+    this.algorithm?.toJwsAlgorithm()?: KmmResult.failure(UnsupportedCryptoException("Unsupported algorithm: ${this::class.simpleName}"))
 
 /** Tries to find a matching JWS algorithm.*/
 fun SpecializedSignatureAlgorithm.toJwsAlgorithm() =
-    this.algorithm.toJwsAlgorithm()
+    this.algorithm?.toJwsAlgorithm()?: KmmResult.failure(UnsupportedCryptoException("Unsupported algorithm: ${this::class.simpleName}"))