+++ indispensable-josef/src/jvmTest/kotlin/at/asitplus/signum/indispensable/josef/ConversionTest.kt
@@ -15,7 +15,7 @@ class ConversionTest : FreeSpec({
     "JWS -> SigAlg -> JWS is stable" - {
         "All" - {
             withData(JwsAlgorithm.entries) {
-                it.algorithm.toJwsAlgorithm() shouldSucceedWith it
+                it.algorithm!!.toJwsAlgorithm() shouldSucceedWith it
             }
         }
         "Specialized SignatureAlgorithm" - {