+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/pki/X509ConversionTests.kt
@@ -13,7 +13,7 @@ class X509ConversionTests : FreeSpec({
     "X509 -> Alg -> X509 is stable" - {
         withData(X509SignatureAlgorithm.entries) {
             it.toX509SignatureAlgorithm() shouldSucceedWith it
-            it.algorithm.toX509SignatureAlgorithm() shouldSucceedWith it
+            it.algorithm!!.toX509SignatureAlgorithm() shouldSucceedWith it
         }
     }
 })