+++ supreme/src/commonTest/kotlin/at/asitplus/signum/supreme/sign/EphemeralSignerCommonTests.kt
@@ -265,7 +265,7 @@ class EphemeralSignerCommonTests : FreeSpec({
                 )
                 val cert = signer.sign(tbsCrt).getOrThrow()
 
-                verifier.verify(cert.tbsCertificate.encodeToDer(), cert.signature) should succeed
+                verifier.verify(cert.tbsCertificate.encodeToDer(), cert.decodedSignature!!) should succeed
 
             }
         }
@@ -317,7 +317,7 @@ class EphemeralSignerCommonTests : FreeSpec({
                 )
                 val cert = signer.sign(tbsCrt).getOrThrow()
 
-                verifier.verify(cert.tbsCertificate.encodeToDer(), cert.signature) should succeed
+                verifier.verify(cert.tbsCertificate.encodeToDer(), cert.decodedSignature!!) should succeed
             }
         }
     }